# openapi.yaml
openapi: 3.1.0
info:
  title: Envelope Sync Relay API
  version: 1.3.0
  description: |
    Zero-knowledge document sync relay for offline-first applications.
    Full specification: docs/envelope-sync-relay/README.md

    **v1.3:** Application registry — `X-ESR-App-Id`, namespace–app binding, operator admin API,
    developer self-service portal, origin verification, pairing `allowedAppIds`.
    See docs/envelope-sync-relay/en/16-APP-REGISTRY.md

    **v1.2:** Multi-document per namespace (`/documents/{documentId}/...`).

    WebSocket notifications (v1.1): wss://{host}/v1/namespaces/{namespaceId}/notifications
    See 13-WEBSOCKET-NOTIFICATIONS.md — not covered by this OpenAPI file.
servers:
  - url: https://sync.example.com/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local development
...
components:
  securitySchemes:
    DeviceToken:
      type: http
      scheme: bearer
      bearerFormat: device_token
    AdminToken:
      type: http
      scheme: bearer
      bearerFormat: admin_api_token
    DeveloperToken:
      type: http
      scheme: bearer
      bearerFormat: developer_jwt
...
  schemas:
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    Limits:
      type: object
      properties:
        freeDeviceLimit:
          type: integer
        purchasedSlots:
          type: integer
        maxDevices:
          type: integer
        activeDevices:
          type: integer
...
  responses:
...
    Error401:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
    Error404:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Error409:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
paths:
...
  /namespaces:
    post:
      tags: [Namespaces]
      summary: Create namespace (first device / host)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - namespaceId
                - namespaceLabel
                - recoveryKeySalt
                - recoveryKeyHash
                - deviceLabel
                - clientDeviceId
              properties:
                namespaceId:
                  type: string
                  format: uuid
                namespaceLabel:
                  type: string
                recoveryKeySalt:
                  type: string
                recoveryKeyHash:
                  type: string
                deviceLabel:
                  type: string
                clientDeviceId:
                  type: string
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  namespaceId:
                    type: string
                  deviceToken:
                    type: string
                  deviceId:
                    type: string
                  limits:
                    $ref: '#/components/schemas/Limits'
        '409':
          $ref: '#/components/responses/Error409'
  /namespaces/{namespaceId}:
    get:
      tags: [Namespaces]
      summary: Get namespace info
      security:
        - DeviceToken: []
      parameters:
        - name: namespaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
...